pyproject.toml
[project]
name = "employee-portal"
version = "0.1.0"
description = "Vault-Tec AI-powered internal knowledge portal"
requires-python = ">=3.11"
dependencies = [
    # Core AI & Retrieval
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "sentence-transformers>=2.2.0",
    "google-generativeai>=0.4.0",
    # Document parsing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.6.0",
    # Backend & Database
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pyjwt>=2.8.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.2.0",
    # Frontend
    "streamlit>=1.32.0",
    "docling>=2.75.0",
    "qdrant-client>=1.17.0",
    "fastembed>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "ragas>=0.1.0",
    "langchain-ollama>=0.1.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
